# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fixly API
  version: 1.0.0
  description: |
    Public + admin HTTP API for Fixly. All responses follow the project's
    envelope contract: `{ "data": ... }` on success and
    `{ "error": { "code": string, "message": string } }` on failure. The
    socket.io surface (`booking:status`, `location:update`) is documented
    under the `Realtime` tag.
servers:
  - url: /api/v1
    description: Same-origin (proxied via Vite in dev, ingress in prod)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string, example: phone must be a valid mobile number }
    Meta:
      type: object
      properties:
        total: { type: integer, example: 42 }
        page: { type: integer, example: 1 }
        pageSize: { type: integer, example: 20 }
...
    Service:
      type: object
      properties:
        id: { type: string, example: elec }
        nameAr: { type: string, example: كهرباء }
        nameEn: { type: string, example: Electricity }
        descriptionAr: { type: string, nullable: true }
        priceJod: { type: string, example: '50.00' }
        durationMin: { type: integer, example: 45 }
        isActive: { type: boolean, example: true }
        category: { type: string, nullable: true, example: HOME_REPAIR }
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum: [AWAITING_PAYMENT, PENDING, CONFIRMED, EN_ROUTE, ARRIVED, IN_PROGRESS, COMPLETED, CANCELLED, DISPUTED, NO_SHOW]
        scheduledAt: { type: string, format: date-time, nullable: true }
        totalJod: { type: string, example: '50.00' }
        service: { $ref: '#/components/schemas/Service' }
    CheckoutSession:
      type: object
      description: >
        Hosted-checkout session for a real PSP (HyperPay COPYandPAY). The client loads
        `scriptUrl?checkoutId=…` and renders a widget for the allowed card `brands`.
      properties:
        checkoutId: { type: string }
        scriptUrl: { type: string, format: uri }
        brands:
          type: array
          items: { type: string }
          example: [VISA, MASTER]
    CreateBookingInput:
      type: object
      required: [serviceId, addressLine, addressLat, addressLng]
      properties:
        serviceId: { type: string }
        addressLine: { type: string, minLength: 3, maxLength: 250 }
        addressLat: { type: number, minimum: -90, maximum: 90 }
        addressLng: { type: number, minimum: -180, maximum: 180 }
        scheduledAt:
          type: string
          format: date-time
          nullable: true
          description: |
            ISO-8601 timestamp. Omit / set null for "now" (within 30 min).
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    DashboardSummary:
      type: object
      properties:
        totalBookings: { type: integer }
        pendingBookings: { type: integer }
        completedBookings: { type: integer }
        totalTechnicians: { type: integer }
        verifiedTechnicians: { type: integer }
        totalRevenueJod: { type: string }
        pendingPayouts: { type: integer }
    Technician:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        phone: { type: string }
        isVerified: { type: boolean }
        rating: { type: number, format: float, nullable: true }
        completedJobs: { type: integer }
    Payout:
      type: object
      properties:
        id: { type: string }
        technicianId: { type: string }
        amountJod: { type: string }
        status: { type: string, enum: [PENDING, PROCESSED, FAILED] }
        requestedAt: { type: string, format: date-time }
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate the session using the httpOnly refresh cookie
      description: >
        Reads the `refresh_token` httpOnly cookie (no request body), rotates it,
        sets a new cookie, and returns a fresh access token. Reuse of a revoked
        cookie revokes the whole session family.
      responses:
        '200':
          description: New access token issued; new refresh cookie set.
          headers:
            Set-Cookie: { schema: { type: string }, description: 'refresh_token=…; HttpOnly; Secure; SameSite=Strict' }
          content:
            application/json:
              schema: { type: object, properties: { data: { type: object, properties: { accessToken: { type: string } } } } }
        '401': { $ref: '#/components/schemas/Error' }
...
  /bookings:
    post:
      tags: [Bookings]
      summary: Create a booking (customer)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBookingInput' }
      responses:
        '201':
          description: >
            Booking created. `checkout` is a hosted-checkout session when a real PSP is
            configured (the booking starts AWAITING_PAYMENT until the customer authorizes
            on the widget); it is null in mock/instant mode (booking is live immediately).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      booking: { $ref: '#/components/schemas/Booking' }
                      checkout:
                        oneOf:
                          - $ref: '#/components/schemas/CheckoutSession'
                          - type: 'null'
        '400': { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/schemas/Error' }
    get:
      tags: [Bookings]
      summary: List the caller's bookings
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: List of bookings.
          content:
            application/json:
              schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Booking' } } } }
  /bookings/{id}:
    get:
      tags: [Bookings]
      summary: Get a single booking (party only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Booking.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Booking' } } }
        '403': { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/schemas/Error' }
...
  /admin/login:
    post:
      tags: [Admin]
      summary: Admin email + password login (sets an httpOnly refresh cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminLoginInput' }
      responses:
        '200':
          description: Access token + admin profile; refresh cookie set.
          headers:
            Set-Cookie: { schema: { type: string }, description: 'admin_refresh_token=…; HttpOnly; Secure; SameSite=Strict' }
          content:
            application/json:
              schema: { type: object, properties: { data: { type: object, properties: { accessToken: { type: string }, admin: { type: object, additionalProperties: true } } } } }
        '401': { $ref: '#/components/schemas/Error' }
  /admin/auth/refresh:
    post:
      tags: [Admin]
      summary: Rotate the admin session via the httpOnly refresh cookie
      responses:
        '200':
          description: New admin access token; new refresh cookie set.
        '401': { $ref: '#/components/schemas/Error' }
...
  /admin/bookings/{id}/refund:
    post:
      tags: [Admin]
      summary: Partial/full refund of a booking's captured payment (audited)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountJod]
              properties:
                amountJod: { type: string, description: 'Positive amount, ≤ outstanding captured, max 3 decimals (fils).' }
      responses:
        '200': { description: Refund applied; updated payment returned. }
        '404': { $ref: '#/components/schemas/Error' }
        '409': { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/schemas/Error' }
  /admin/stats:
    get:
      tags: [Admin]
      summary: Aggregate dashboard metrics
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Metrics.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/DashboardSummary' } } }
  /admin/technicians:
    get:
      tags: [Admin]
      summary: List technicians (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: verified
          schema: { type: boolean }
        - in: query
          name: search
          description: Matches phone (contains) or name (case-insensitive contains).
          schema: { type: string }
      responses:
        '200':
          description: Paginated technicians.
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Technician' } }
                  - $ref: '#/components/schemas/Meta'
  /admin/technicians/{id}/verify:
    post:
      tags: [Admin]
      summary: Verify a technician (idempotent)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Verification state.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Technician' } } }
  /admin/bookings:
    get:
      tags: [Admin]
      summary: List bookings (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Paginated bookings.
  /admin/bookings.csv:
    get:
      tags: [Admin]
      summary: Bookings CSV export (OPS) — same status filter as the list endpoint, capped at 5000 rows
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [AWAITING_PAYMENT, PENDING, CONFIRMED, EN_ROUTE, ARRIVED, IN_PROGRESS, COMPLETED, CANCELLED, DISPUTED, NO_SHOW] }
      responses:
        '200': { description: CSV attachment (id, status, customer, technician, service, amounts, timestamps). }
  /admin/bookings/{id}:
    get:
      tags: [Admin]
      summary: Full booking detail for the admin drawer — booking + status history + additional work + payment (OPS)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Detail. Access is audited.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      booking: { $ref: '#/components/schemas/Booking' }
                      statusHistory:
                        type: array
                        items:
                          type: object
                          properties:
                            status: { type: string }
                            changedAt: { type: string, format: date-time }
                      additionalWork:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: string, format: uuid }
                            description: { type: string }
                            amountJod: { type: string }
                            status: { type: string }
                            createdAt: { type: string, format: date-time }
                      payment:
                        type: object
                        nullable: true
        '404': { $ref: '#/components/schemas/Error' }
  /admin/payouts:
    get:
      tags: [Admin]
      summary: List payouts (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, PROCESSED, FAILED] }
      responses:
        '200':
          description: Paginated payouts.
  /admin/payouts/{id}/process:
    post:
      tags: [Admin]
      summary: Mark a payout as processed
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Payout state.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Payout' } } }
        '409': { $ref: '#/components/schemas/Error' }
  /admin/customers:
    get:
      tags: [Admin]
      summary: List customers (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: search
          description: Matches phone (contains) or name (case-insensitive contains).
          schema: { type: string }
      responses:
        '200':
          description: Paginated customers.

  # ---- Reviews (public) ----
...